+++ attributes/banner.rb
@@ -1,4 +1,4 @@
 default[:sys][:banner][:header] = String.new
 default[:sys][:banner][:message] = String.new
 default[:sys][:banner][:footer] = String.new
-default[:sys][:banner][:info] = true
+default[:sys][:banner][:info] = false