+++ attributes/sudo.rb
@@ -1,0 +1,1 @@
+default[:sys][:sudo] = Mash.new