+++ attributes/ssh.rb
@@ -1,2 +1,2 @@
 default[:sys][:ssh][:authorize] = Mash.new
-default[:sys][:sshd][:config] = Hash.new 
+default[:sys][:sshd][:config] = Hash.new