+++ definitions/sys_sudo.rb
@@ -21,8 +21,8 @@ define :sys_sudo, :users => Hash.new, :hosts => Hash.new, :commands => Hash.new,
     owner 'root'
     group 'root'
     mode 0440
-    variables( 
-      :name => name, 
+    variables(
+      :name => name,
       :users => params[:users],
       :hosts => params[:hosts],
       :commands => params[:commands],