+++ CHANGELOG.md
@@ -13,5 +13,5 @@
 * Deploy cgroups with attributes in `node.sys.cgroups`.
 * Alter Grub boot configuration attributes in `node.sys.boot`. 
 * System login banner with attributes in `node.sys.banner`,
-* Load Linux kernel modules with `linux_module`.
+* Load Linux kernel modules with `sys_module`.
 * Reboot/Shutdown node with `sys_shutdown`.