+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.8.0
 
+* Configuration of Sudo with `node.sys.sudo`.
 * Configuration of the SSH daemon with `node.sshd.config` and management
   of authorized keys with `node.sys.ssh.authorize`.
 * Support the configuration of network interfaces, VLANs and bridges with
@@ -12,5 +13,5 @@
 * Deploy cgroups with attributes in `node.sys.cgroups`.
 * Alter Grub boot configuration attributes in `node.sys.boot`. 
 * System login banner with attributes in `node.sys.banner`,
-* Load Linux kernel modules with `linux_module`.
+* Load Linux kernel modules with `sys_module`.
 * Reboot/Shutdown node with `sys_shutdown`.