+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-# next version
+# 0.10.0
 
+* Add LDAP/Kerberos support.
 * Info banner for interactive shells only.
 
 # 0.9.0